Split out teacher data management functional tests into their own modules
